api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["gutrapp <trappgustavo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg = "^3.1.10"
python-decouple = "^3.8"
django-filter = "^23.3"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.9"
datetime = "^5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
